+++ apps/playground-web/src/lib/constants.ts
@@ -1,0 +1,3 @@
+export const metadataBase = process.env.VERCEL_ENV
+  ? new URL("https://playground.thirdweb.com")
+  : undefined;