+++ legacy_packages/chains/chains/336666.ts
@@ -6,13 +6,7 @@ export default {
     {
       "name": "UPchain Mainnet Explorer",
       "url": "https://explorer.uniport.network",
-      "standard": "EIP3091",
-      "icon": {
-        "url": "ipfs://QmNMqcPyQnaaFfUDsmsnu2T6agZfoWij3zFYebCNUPkfAH",
-        "width": 320,
-        "height": 320,
-        "format": "jpg"
-      }
+      "standard": "EIP3091"
     }
   ],
   "faucets": [],
@@ -21,12 +15,6 @@ export default {
       "name": "EIP155"
     }
   ],
-  "icon": {
-    "url": "ipfs://QmNMqcPyQnaaFfUDsmsnu2T6agZfoWij3zFYebCNUPkfAH",
-    "width": 320,
-    "height": 320,
-    "format": "jpg"
-  },
   "infoURL": "https://uniport.network",
   "name": "UPchain Mainnet",
   "nativeCurrency": {