+++ legacy_packages/chains/chains/7000.ts
@@ -10,12 +10,6 @@ export default {
     }
   ],
   "faucets": [],
-  "icon": {
-    "url": "ipfs://QmP4Gnf4Lkp8q5LQVePNjAWxSqrw8vU2JAf7amcFz4vEUy",
-    "width": 712,
-    "height": 712,
-    "format": "png"
-  },
   "infoURL": "https://zetachain.com/docs/",
   "name": "ZetaChain Mainnet",
   "nativeCurrency": {